refactor(store): extract store/utility repository from database.py
Move 10 EVU/VNB utility-client funcs (accounts, magic-link tokens, status,
API keys, stats) into store/utility.py, re-exported from database.py so
db.<fn>() keeps working. Completes Candidate #4 wave (ranking, profile,
billing, email_queue, utility). No behavior change; full suite + ruff green.

Closes #79
